change wording from quarter to quartile (grammar mistake) and statistics
the previous statistics gave the mean, which I don't think is useful as the 4th quartile will be completely randomised and will likely lower the mean.

in addition, calculating the average of the higher and lower fitnesses would make more sense from the 1st to 3rd (not 4th) quartile, as it should generally ignore bad-performing models from the 4th quarter (as they are completely random).