add extra channel for input tensor indicating bounds of screen
the previous method of showing no bullets in out-of-bound spots will mislead models to believe that inaccessible areas are accessible and are safe from being hit by bullets.

this is especially problematic as the model doesn't have a concept of time (information from a single point only), so it would be near-impossible for it to know certain areas were inaccessible without somehow memorising the bounds

this extra channel indicates with a 0 or 1 if the pixel is out or in bounds respectively, which should (hopefully) allow the models to learn overtime to avoid those areas and try alternative routes to dodge bullets.