implement multiprocessing for Q-learning
this is a huge commit, not helped by Windows not supported pickling objects, which led me to create functions for objects to convert to lists of parameters that could be used to re-initialise them.

by all means, this is not an elegant solution, and is rather a workaround to use multiprocessing.