Model: processes can process multiple frames without restarting
this commit was made to prevent the issue of restarting processes every frame, which should speed up execution.

this, however, is more focused towards Windows, which - unlike macOS and Linux - requires **every library** to be reloaded for **every process**, **every frame**, which took up so much time that single-threaded performance was faster.