Game: change screen bounds to cover size of window, make bullets smaller in screen
important note: following the previous commits, the neural network has produced extremely promising results, having increased confidence over training and increased reward.

the input tensor from get_screen() used to cover up to two-thirds of the width and height of the window, which I believe could be used to improve the model by increasing it, essentially increasing its view distance.

I changed the input tensor to cover the full width and height of the window, meaning that if at the centre, the model can view the whole window.

the way the bullets were rendered on screen was as a 9x9 square with no range of intensity, despite the fact that the centre was the bigger danger, so now the edges have a lesser intensity of 0.75. in addition, the 9x9 squares were simply too large, so have been reduced to 4x4 squares instead.