Common: increase GAME_FPS, FRAMES_PER_ACTION, NUM_BULLETS and decrease BULLET_HONE_SPEED
increased GAME_FPS and FRAMES_PER_ACTION should result in smoother moves, as the low GAME_FPS set led to very unsmooth gameplay, and the worry about long training times seems to not be an issue.

increased bullets should benefit the model in dodging bullets in high-intensity situations, which - along with slow bullet speed (although this is more Touhou-centric) - are staples of the danmaku that this model should eventually run on.