Model: skip rest of loop if no bullet or player touches bullet
before, if there was no bullet on screen, the already-made training data in exp_inps and exp_outs would be removed, wasting time. now, once Game is updated with action from Q-learning agent, the rest of the code (estimating confidence, printinting stats) is skipped for the next iteration.

in addition, the rest of the loop is skipped if the Q-value of the chosen action is too low (only possible if player touches bullet), as this training data shouldn't be replicated by the model as it shows what not to do.