Train: reduce network size to prevent overfitting
while a large network for training is generally ideal, my training results show a steep increase in fitness, followed by a slow decrease and lack of change of fitness that is paired with a much steeper decrease in traininng error.

I concluded that this was the model overfitting, as it performed better and better on the training data, but similar if not worse when validating, which can be mitigated through small networks.