Game + Model: more efficient Q-agent algorithm by skipping unncecessary frames and reducing calls to multiprocessing dictionary
imagine that the Q-agent has found an action with an "ideal" (more than -9e90, so bullet didn't collide with the player) Q-value, but continues to look at actions as well as the possible actions (action_new) within that action despite having an "unideal" Q-value. we can save time by completely skipping looking at the possible actions and just give the Q-value a low score (-9e90), because it will be ignored later, anyway.

in Action_Update() (and by extension, Sim_Update()), the score at any frame may become very low (less than -9e90) as a result of colliding with a bullet, but the loop continues despite the low score being known and now knowing that the whole action will be skipped anyways because of said low score. just like the first example, we can just skip the rest of the frames too.

lastly, setting values to the multiprocessing manager's dictionaries are quite expensive, so instead of assigning each action an index, we assign every action in each process an index. this approach may not make much sense, but it does save time as well.

to put these time saves in perspective, the first point saved on average 13.0%, the second point 0.047% and the third point 12.6% to save in total 24.0% of processing time, which is of course one of the main aspects of building the dataset and the project as a whole.