substitute large numbers (mostly 9e99) for float('inf')
throughout the project, it was assumed that infinity could not be represented in standard, built-in Python libraries, but recently I have discovered that it can be.

previously, using large numbers posed an issue as when added with other numbers they could change from the specified number, so when checking if it was large, I would compare with a smaller but still large number, which was a very "got the work done" workaround and a very poor solution.

 using infinity not only increases clarity as to what the large numbers mean, but incrementing and decrementing is _guaranteed_ to make no changes to infinity.