switch penalty system from hitboxes to bullet distance to player
the fundamental issue with hitboxes is that they assume that any bullet in the hitbox has an equal penalty regardless of its proximity to the player.

this is why two hitboxes ("touching", grazing) are used, but ideally there should be more, smaller hitboxes that account for all quantities of a penalty.

additionally, the largest problem with the Q-agent is that as long as there are no bullets within the player's hitboxes, they will essentially go in random directions as all moves won't cause any penalty, despite this usually putting it into a far more dangerous situation that the AI model will replicate. changing constants in Common (FRAMES_PER_ACTION, TRAIN_FPS, DISCOUNT_RATE, GRAZE_SIZE, TOUCH_SIZE) helped _reduce_ this problem, **but could never solve it.**

to solve this problem, for every bullet, a penalty is calculated through a mathematical function that exponentially increases the given penalty as a bullet's distance to the player decreases, allowing the Q-agent to pick the most optimal movement w/o blind spots and to detect immediate danger as a bullet comes in close distance to the player.