Game: fix crash when all actions are -inf (bullet collisions)
the reason is quite hard to explain, but I'll try my best to do so:

the Q-value is the change in score from before to after an action.

if building the dataset (BUILD\_DL), when a bullet collides with the player, the penalty was -9e99 and is now -inf - both huge negative numbers - to deter any justification of colliding with a bullet in the training dataset.

the issue is if the previous Game score was -inf (if the player _had_ to collide with a bullet) and the action resulted in - say - a penalty of -1, **the new Game score would still be -inf.**

this was likely an issue for a while now, but when the penalty was still -9e99, the Q-value would've been 0 and still a number. **but when using -inf instead, the Q-value became NaN, causing the random crashes.

this commit returns -inf to QAgent - instead resulting in a Q-value of -inf - but does not actually give a -inf penalty to Game, so all future Q-values will be correct. in addition, the q_fitness will actually match the Game score given if the Q-agent was actually playing the game, so that data will be fully useful, as well.