Game: use absolute rather than relative bullet position in get_screen()
relative bullet position - where the player is always at the centre of the input tensor - was the original method, based on the idea that I wouldn't use a convolutional neural network (which I am).

an absolute bullet position should make **little to no difference in neural network performance**, with the added benefit that **it can naturally comprehend the bounds of the window**: one of the largest problems with this project so far.